requirements.txt
amqp==2.4.0
backcall==0.1.0
billiard==3.5.0.5
cachetools==3.1.0
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
djangorestframework==3.9.1
google-auth==1.6.2
gunicorn==19.9.0
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
kombu==4.2.2.post1
Markdown==3.0.1
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.3.1
PyMySQL==0.9.3
python-dateutil==2.7.5
pytz==2018.9
redis==3.1.0
requests==2.21.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.2.17
traitlets==4.3.2
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
